requirements.txt
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dateutils==0.6.12
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
filelock==3.12.2
fonttools==4.41.1
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==10.0.0
preshed==3.0.8
pydantic==1.10.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-splitter==1.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
sympy==1.12
thinc==8.1.10
thinc-apple-ops==0.1.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
zipp==3.16.2
